refactor: improve code structure and readability across controllers, models, and tests
* add: frozen string literal comment to enhance performance
* refactor: streamline require_admin method in Admin controllers for clarity
* refactor: update user role validation and helper modules for consistency
* add: new concerns for user activity tracking and streak management
* refactor: simplify user model methods and remove unused code
* update: modify test files to include frozen string literal and improve structure
