+++ app/helpers/admin/dashboard_helper.rb
@@ -1,2 +1,6 @@
-module Admin::DashboardHelper
+# frozen_string_literal: true
+
+module Admin
+  module DashboardHelper
+  end
 end
