+++ app/helpers/admin/users_helper.rb
@@ -1,2 +1,6 @@
-module Admin::UsersHelper
+# frozen_string_literal: true
+
+module Admin
+  module UsersHelper
+  end
 end
