+++ db/migrate/20250508090046_add_role_to_users.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class AddRoleToUsers < ActiveRecord::Migration[6.1]
   def change
     add_column :users, :role, :string, default: 'user'
