+++ db/migrate/20250508091937_create_metrics.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class CreateMetrics < ActiveRecord::Migration[7.0]
   def change
     create_table :metrics do |t|
