+++ test/controllers/admin/users_controller_test.rb
@@ -1,18 +1,22 @@
-require "test_helper"
+# frozen_string_literal: true
 
-class Admin::UsersControllerTest < ActionDispatch::IntegrationTest
-  test "should get edit" do
-    get admin_users_edit_url
-    assert_response :success
-  end
+require 'test_helper'
 
-  test "should get update" do
-    get admin_users_update_url
-    assert_response :success
-  end
+module Admin
+  class UsersControllerTest < ActionDispatch::IntegrationTest
+    test 'should get edit' do
+      get admin_users_edit_url
+      assert_response :success
+    end
 
-  test "should get destroy" do
-    get admin_users_destroy_url
-    assert_response :success
+    test 'should get update' do
+      get admin_users_update_url
+      assert_response :success
+    end
+
+    test 'should get destroy' do
+      get admin_users_destroy_url
+      assert_response :success
+    end
   end
 end
