+++ test/models/metric_test.rb
@@ -1,4 +1,6 @@
-require "test_helper"
+# frozen_string_literal: true
+
+require 'test_helper'
 
 class MetricTest < ActiveSupport::TestCase
   # test "the truth" do
